+++ test/gplPalette/test.py
@@ -11,7 +11,7 @@ import sys
 from pathlib import Path
 PROJECTDIR = str(Path(__file__).resolve().parent.parent)
 sys.path.insert(0, os.path.dirname(PROJECTDIR))
-from gimpformats_unofficial import GimpGplPalette
+from gimpformats import GimpGplPalette
 
 __HERE__ = os.path.abspath(__file__).rsplit(os.sep, 1)[0] + os.sep