+++ test/simpleXcfRead/test.py
@@ -12,7 +12,7 @@ from pathlib import Path
 PROJECTDIR = str(Path(__file__).resolve().parent.parent)
 sys.path.insert(0, os.path.dirname(PROJECTDIR))
 from imgcompare import imgcompare
-from gimpformats_unofficial import GimpDocument
+from gimpformats import GimpDocument
 
 __HERE__ = os.path.abspath(__file__).rsplit(os.sep, 1)[0] + os.sep
 
@@ -30,8 +30,15 @@ class Test(unittest.TestCase):
 	def testImage(self):
 		""" test an image """
 		self.dut.load(__HERE__ + 'one_layer_with_transparency.xcf')
-		self.dut.save(__HERE__ + 'actualOutput.png')
+		#self.dut.save(__HERE__ + 'actualOutput.png')
+		self.dut.image.save(__HERE__ + 'actualOutput.png')
 		assert imgcompare.is_equal(self.dut.image, __HERE__ + 'desiredOutput.png', tolerance=1)
+
+	def testComplexImage(self):
+		""" test an image """
+		self.dut.load(__HERE__ + 'base24.xcf')
+		self.dut.image.save(__HERE__ + 'base24.png')
+		assert imgcompare.is_equal(self.dut.image, __HERE__ + 'expectedNoHidden.png', tolerance=1)
 
 
 def testSuite():
@@ -40,6 +47,7 @@ def testSuite():
 	"""
 	varTestSuite = unittest.TestSuite()
 	varTestSuite.addTest(Test("testImage"))
+	varTestSuite.addTest(Test("testComplexImage"))
 	return varTestSuite