+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.1 - 2020/05/03
+- Attempted some fixes to gimpImageInternals.py
+- Fix gimpIOBase.py
+- Added image compositor and added a new test for a 'complex' image
+- Now going to release under gimpformats
+
 ## 2020.0.2 - 2020/05/02
 - Added documentation
 - Project icon :smile: